+++ controllers/util/util.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	eraserv1 "github.com/Azure/eraser/api/v1"
-	"github.com/Azure/eraser/pkg/config"
+	"github.com/Azure/eraser/api/v1/config"
 	"github.com/Azure/eraser/pkg/utils"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
